requirements.txt
# Universal Web Scraper Dependencies
# Core scraping and crawling
playwright==1.40.0
scrapy==2.11.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
urllib3>=1.26.14,<2.1.0

# Content processing
readability-lxml==0.8.1
html2text==2020.1.16
markdown==3.5.1

# Media processing
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
pydub==0.25.1
moviepy==1.0.3

# AI/ML libraries
torch>=2.1.0
transformers>=4.35.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3

# Vector database
chromadb==0.4.18
qdrant-client==1.6.4

# NLP processing
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3

# Async processing
celery==5.3.4
redis==5.0.1
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Data formats
pandas==2.1.3
pyarrow==14.0.1
orjson==3.9.10
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1